加入aes加密功能，支持自定义Application类的加固

+++ jni/byte_load.cpp
@@ -1,8 +1,17 @@
+/**
+ * @brief 
+ * 
+ * @file byte_load.cpp
+ * @author your name
+ * @date 2018-05-08
+ */
 #include "byte_load.h"
 #include <jni.h>
 #include <android/log.h>
 #include <errno.h>
 #include <assert.h>
+#include <stdio.h>
+#include <string.h>
 #include <string>
 #include <cstdlib>
 #include <sys/stat.h>
@@ -11,6 +20,7 @@
 #include <sys/mman.h>
 #include <vector>
 #include <bits/unique_ptr.h>
+
 #include "packer.h"
 #include "dex_header.h"
 #include "elfGotHook/logger.h"
@@ -23,58 +33,56 @@ using namespace std;
 //
 void *mem_loadDex_byte19(void *arthandler, const char *base, size_t size)
 {
-    std::string                 location = "";
-    std::string                 err_msg;
+    std::string location = "";
+    std::string err_msg;
     org1_artDexFileOpenMemory19 func = (org1_artDexFileOpenMemory19)dlsym(arthandler,
                                                                           "_ZN3art7DexFile10OpenMemoryEPKhjRKSsjPNS_6MemMapE");
-    if (!func) {
+    if (!func)
+    {
         return NULL;
     }
     const Header *dex_header = reinterpret_cast<const Header *>(base);
-    void         *value      = func((const unsigned char *)base,
-                                    size,
-                                    location,
-                                    dex_header->checksum_,
-                                    NULL);
-    if (!value) {
+    void *value = func((const unsigned char *)base,
+                       size,
+                       location,
+                       dex_header->checksum_,
+                       NULL);
+    if (!value)
+    {
         LOGE("[-]sdk_int:%d dlsym openMemory failed:%s", g_sdk_int, dlerror());
         return NULL;
     }
     return value;
 }
-
 
 void *mem_loadDex_byte21(void *artHandle, const char *base,
                          size_t size)
 {
     std::string location = "";
-
     std::string err_msg;
-
     org_artDexFileOpenMemory21 func = (org_artDexFileOpenMemory21)dlsym(artHandle,
                                                                         "_ZN3art7DexFile10OpenMemoryEPKhjRKNSt3__112basic_stringIcNS3_11char_traitsIcEENS3_9allocatorIcEEEEjPNS_6MemMapEPS9_");
 
-    if (!func) {
+    if (!func)
+    {
         return NULL;
     }
-
     const Header *dex_header = reinterpret_cast<const Header *>(base);
-    void         *value      = func((const unsigned char *)base,
-                                    size,
-                                    location,
-                                    dex_header->checksum_,
-                                    NULL,
-                                    &err_msg);
+    void *value = func((const unsigned char *)base,
+                       size,
+                       location,
+                       dex_header->checksum_,
+                       NULL,
+                       &err_msg);
 
-    if (!value) {
+    if (!value)
+    {
         LOGE("[-]sdk_int:%d dlsym openMemory failed:%s", g_sdk_int, dlerror());
         return NULL;
     }
 
-
     return value;
 }
-
 
 void *mem_loadDex_byte22(void *artHandle, const char *base, size_t size)
 {
@@ -84,7 +92,8 @@ void *mem_loadDex_byte22(void *artHandle, const char *base, size_t size)
     org_artDexFileOpenMemory22 func = (org_artDexFileOpenMemory22)dlsym(artHandle,
                                                                         "_ZN3art7DexFile10OpenMemoryEPKhjRKNSt3__112basic_stringIcNS3_11char_traitsIcEENS3_9allocatorIcEEEEjPNS_6MemMapEPKNS_7OatFileEPS9_");
 
-    if (!func) {
+    if (!func)
+    {
         LOGE("[-]sdk_int:%d dlsym openMemory failed:%s", g_sdk_int, dlerror());
         return NULL;
     }
@@ -98,14 +107,14 @@ void *mem_loadDex_byte22(void *artHandle, const char *base, size_t size)
                        NULL,
                        &err_msg);
 
-    if (!value) {
+    if (!value)
+    {
         LOGE("[-]call artDexFileOpenMemory22 failed");
         return NULL;
     }
     LOGD("[+]openMemory value:%p", value);
     return value;
 } // mem_loadDex_byte22
-
 
 /**
  * [mem_loadDex_byte23 description]
@@ -115,27 +124,31 @@ void *mem_loadDex_byte22(void *artHandle, const char *base, size_t size)
  */
 void *mem_loadDex_byte23(void *artHandle, const char *base, size_t size)
 {
+
     std::string location = "";
     std::string err_msg;
-    void        *retcookie = malloc(0x78);
-    memset(retcookie, 0, 0x78);
+
+    void *retcookie = malloc(0x78);
+    memset(retcookie, 0, (size_t)0x78);
     org_artDexFileOpenMemory23 func = (org_artDexFileOpenMemory23)dlsym(artHandle,
                                                                         "_ZN3art7DexFile10OpenMemoryEPKhjRKNSt3__112basic_stringIcNS3_11char_traitsIcEENS3_9allocatorIcEEEEjPNS_6MemMapEPKNS_10OatDexFileEPS9_");
 
-    if (!func) {
+    if (!func)
+    {
         LOGE("[-]sdk_int:%d dlsym openMemory failed:%s", g_sdk_int, dlerror());
-
         return NULL;
     }
+
     const Header *dex_header = reinterpret_cast<const Header *>(base);
-    void         *value      = func(retcookie,
-                                    (const unsigned char *)base,
-                                    size,
-                                    location,
-                                    dex_header->checksum_,
-                                    NULL,
-                                    NULL,
-                                    &err_msg);
+    void *value = func(retcookie,
+                       (const unsigned char *)base,
+                       size,
+                       location,
+                       dex_header->checksum_,
+                       NULL,
+                       NULL,
+                       &err_msg);
+
     void *a = retcookie;
 
     // 返回的value等于retcookie ,*(int*)retcookie存储了加载dex的cookie
@@ -144,42 +157,43 @@ void *mem_loadDex_byte23(void *artHandle, const char *base, size_t size)
     return (void *)(*(jlong *)retcookie);
 }
 
-
 void *mem_loadDex_byte24(void *artHandle, const char *base, size_t size)
 {
     std::string location = "";
     std::string err_msg;
-    void        *retcookie = malloc(0x78);
+    void *retcookie = malloc(0x78);
     memset(retcookie, 0, 0x78);
 
     // #define SEARCH_SYMBOL_Nougat																																			_ZN3art7DexFile10OpenMemoryEPKhjRKNSt3__112basic_stringIcNS3_11char_traitsIcEENS3_9allocatorIcEEEEjPNS_6MemMapEPKNS_10OatDexFileEPS9_
     org_artDexFileOpenMemory23 func = (org_artDexFileOpenMemory23)dlsym(artHandle,
                                                                         "_ZN3art7DexFile10OpenMemoryEPKhjRKNSt3__112basic_stringIcNS3_11char_traitsIcEENS3_9allocatorIcEEEEjPNS_6MemMapEPKNS_10OatDexFileEPS9_");
 
-    if (!func) {
+    if (!func)
+    {
         LOGE("[-]sdk_int:%d dlsym openMemory failed:%s", g_sdk_int, dlerror());
 #ifndef SEARCH_SYMBOL_Nougat
         return NULL;
-#else         // ifndef SEARCH_SYMBOL_Nougat
+#else  // ifndef SEARCH_SYMBOL_Nougat
         LOGD("[+]try search symbol from elf file");
         func = (org_artDexFileOpenMemory23)get_addr_symbol("/system/lib/libart.so",
                                                            "_ZN3art7DexFile10OpenMemoryEPKhjRKNSt3__112basic_stringIcNS3_11char_traitsIcEENS3_9allocatorIcEEEEjPNS_6MemMapEPKNS_10OatDexFileEPS9_");
 
-        if (!func) {
+        if (!func)
+        {
             LOGE("[-]search symbol openMemory uniptr failed");
             return NULL;
         }
-#endif         // ifndef SEARCH_SYMBOL_Nougat
+#endif // ifndef SEARCH_SYMBOL_Nougat
     }
     const Header *dex_header = reinterpret_cast<const Header *>(base);
-    void         *value      = func(retcookie,
-                                    (const unsigned char *)base,
-                                    size,
-                                    location,
-                                    dex_header->checksum_,
-                                    NULL,
-                                    NULL,
-                                    &err_msg);
+    void *value = func(retcookie,
+                       (const unsigned char *)base,
+                       size,
+                       location,
+                       dex_header->checksum_,
+                       NULL,
+                       NULL,
+                       &err_msg);
     void *a = retcookie;
 
     // LOGD改变了retcookie？？ 所以先用a备份
@@ -190,7 +204,6 @@ void *mem_loadDex_byte24(void *artHandle, const char *base, size_t size)
 
     return (void *)(*(jlong *)retcookie);
 } // mem_loadDex_byte24
-
 
 // For Andoird oreo 8.0 and 8.1
 // Reserved
